Add two other queries mentioned in talk description
In the talk description [1], I mentioned three example queries. One of
them (largest cities with female mayors) is already explained in the
slides; the other two are much too long for that, but we can at least
put them on the final slide.

One of the short URLs includes a dollar sign, which needs to be escaped.
I tried a bit to tweak the \wwiki command so it would override the
catcode of $ from math shift to letter or other, but it didn’t work – I
think it’s because Beamer reads the whole frame content at once, already
tokenizing it, before \wwiki has a chance to run. (Even \catcode`$=11
directly in the \wwiki argument didn’t work – it only worked when I put
it between frames.)

[1]: https://cfp.verschwoerhaus.de/36c3/talk/AMPBFW/